Move files around to `jws` package
It makes more sense to tightly couple the verifiers and signers
now that we're changing the API to a lookup based approach
