+++ jws/option.go
@@ -1,7 +1,6 @@
 package jws
 
 import (
-	"github.com/lestrrat-go/jwx/jws/sign"
 	"github.com/lestrrat-go/option"
 )
 
@@ -10,7 +9,7 @@ type Option = option.Interface
 type identPayloadSigner struct{}
 type identHeaders struct{}
 
-func WithSigner(signer sign.Signer, key interface{}, public, protected Headers) Option {
+func WithSigner(signer Signer, key interface{}, public, protected Headers) Option {
 	return option.New(identPayloadSigner{}, &payloadSigner{
 		signer:    signer,
 		key:       key,
