+++ jws/signer.go
@@ -1,1 +1,54 @@
 package jws
+
+import (
+	"github.com/lestrrat-go/jwx/jwa"
+	"github.com/pkg/errors"
+)
+
+type SignerFactory interface {
+	Create(jwa.SignatureAlgorithm) (Signer, error)
+}
+type SignerFactoryFn func(jwa.SignatureAlgorithm) (Signer, error)
+
+func (fn SignerFactoryFn) Create(sig jwa.SignatureAlgorithm) (Signer, error) {
+	return fn(sig)
+}
+
+var signerDB map[jwa.SignatureAlgorithm]SignerFactory
+
+// RegisterSigner is used to register a factory object that creates
+// Signer objects based on the given algorithm.
+//
+// For example, if you would like to provide a custom signer for
+// jwa.EdDSA, use this function to register a `SignerFactory`
+// (probably in your `init()`)
+func RegisterSigner(alg jwa.SignatureAlgorithm, f SignerFactory) {
+	signerDB[alg] = f
+}
+
+func init() {
+	signerDB = make(map[jwa.SignatureAlgorithm]SignerFactory)
+
+	for _, alg := range []jwa.SignatureAlgorithm{jwa.RS256, jwa.RS384, jwa.RS512, jwa.PS256, jwa.PS384, jwa.PS512} {
+		RegisterSigner(alg, SignerFactoryFn(newRSASigner))
+	}
+
+	for _, alg := range []jwa.SignatureAlgorithm{jwa.ES256, jwa.ES384, jwa.ES512} {
+		RegisterSigner(alg, SignerFactoryFn(newECDSASigner))
+	}
+
+	for _, alg := range []jwa.SignatureAlgorithm{jwa.HS256, jwa.HS384, jwa.HS512} {
+		RegisterSigner(alg, SignerFactoryFn(newHMACSigner))
+	}
+
+	RegisterSigner(jwa.EdDSA, SignerFactoryFn(newEdDSASigner))
+}
+
+// NewSigner creates a signer that signs payloads using the given signature algorithm.
+func NewSigner(alg jwa.SignatureAlgorithm) (Signer, error) {
+	f, ok := signerDB[alg]
+	if ok {
+		return f.Create(alg)
+	}
+	return nil, errors.Errorf(`unsupported signature algorithm %s`, alg)
+}
